postgresql output: batch inserts to the same table/column set
Batches are created for every (table, columns) set as we may have metrics
with different set of tags & fields in the same batch.  This speeds up
inserts quite a bit by reducing the number of separate insert statements.

Grouping inserts requires columns to appear in the same order, so we now
sort them for both tags & fields.  This code could use a bit of tidying.