Add a hold_time parameter to the merge aggregator.
The purpose of this parameter is to ensure that points which can be merged don't get split across batches due to a flush occurring in the middle of them.

This is accomplished by ensuring new points, which are not merged into prior points, are held for a minimum amount of time, allowing latter points to be merged in.