outputs.postgresql: Drop metrics with missing tag columns in DB
When schema updates are disabled, drop any metrics with tags that aren't in the database. We don't want to just throw an error as the rest of the metrics might be fine. We also don't want to insert without that tag as this can result in a collision with other metrics.